Optimize how items are added to Reading Quantified
Although the server checks for unique Trello IDs, the client also checks for valid data (e.g. valid dates for book objects). It also queries for all items from the server first before attempting to add an item. This reduces the overall number of query attempts that are made to the server.